Move details into details folder
