--- tests/beman/big_int/todo.test.cpp
@@ -1,10 +1,0 @@
-// SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
-
-#include <beman/big_int/todo.hpp>
-
-#include <gtest/gtest.h>
-
-TEST(TodoTest, todo) {
-    const bool todo = true;
-    EXPECT_TRUE(todo);
-}
